Move ACC files to their own group
