+++ AzureCommunicationUI/sdk/AzureCommunicationUICalling/Sources/Service/Calling/Extension/AzureCommunicationCalling/LocalVideoStreamExtension.swift
@@ -6,19 +6,6 @@
 import AzureCommunicationCalling
 import Foundation
 
-extension CameraFacing {
-    func toCameraDevice() -> CameraDevice {
-        switch self {
-        case .front:
-            return .front
-        case .back:
-            return .back
-        default:
-            return .front
-        }
-    }
-}
-
 extension LocalVideoStream {
     static func nativeAccMediaStreamType(type: MediaStreamType)
     -> AzureCommunicationCalling.MediaStreamType {
