+++ AzureCommunicationUI/sdk/AzureCommunicationUICalling/Sources/Service/Calling/Extension/AzureCommunicationCalling/RemoteParticipantExtension.swift
